Extract common code into AbstractTaskBuilder
